feat: Add getCompaniesService to retrieve list of companies
This commit adds the `getCompaniesService` function to the `companyServices.ts` file. This function retrieves a list of companies from the database using the `getCompaniesRepository` function. The list of companies is returned along with the count of companies. If an error occurs while getting the companies, an error message is thrown. This service function will be used by the `getCompaniesController` to fetch the companies and return them as a JSON response.