feat: Add findCompanyByIdRepository function
This commit adds the `findCompanyByIdRepository` function to the `companyRepositories.ts` file. This function retrieves a company from the database based on its ID. It uses the `companyId` parameter to query the database and returns the company details as a `CompanyResponseDTO` object. If the company does not exist, an error is thrown.