feat: Add createProductController function
The `createProductController` function in `productController.ts` has been added to handle the creation of products. This change allows for the creation of new products by accepting product data from the request body and calling the `createProductService` function. The function returns the newly created product as a JSON response with a status code of 201. In case of an error, a 500 status code is returned along with an error message. This enhancement improves the functionality of the application by providing a complete CRUD (Create, Read, Update, Delete) capability for products.