feat: Add createProductService function
The `createProductService` function in `productService.ts` has been added to handle the creation of products. This change allows for the creation of new products by accepting product data and calling the `createProductRepository` function. The function returns the newly created product. This enhancement improves the functionality of the application by providing a complete CRUD (Create, Read, Update, Delete) capability for products.