feat: Add getProductsController function
The `getProductsController` function in `productController.ts` has been added to handle the retrieval of products. This change allows for fetching all products by calling the `getProductsService` function and returning them as a JSON response with a status code of 200. In case of an error, a 500 status code is returned along with an error message. This enhancement improves the functionality of the application by providing a complete CRUD (Create, Read, Update, Delete) capability for products.