feat: Add getProductsRepository function
The `getProductsRepository` function in `productRepository.ts` has been added to handle the retrieval of products. This change allows for fetching all products by querying the database collection and returning them as an array of `ProductResponseDTO`. This enhancement improves the functionality of the application by providing the ability to retrieve products based on the `companyId`.