feat: Add getProductsController and getProductsRepository functions
This commit adds the `getProductsController` and `getProductsRepository` functions to the `productController.ts` and `productRepository.ts` files respectively. These functions enable the retrieval of products from the database and the handling of the corresponding HTTP request. This enhancement improves the functionality of the application by providing the ability to fetch all products and display them to the user.