feat: Add getProductsService function
The `getProductsService` function in `productService.ts` has been added to handle the retrieval of products. This change allows for fetching all products by calling the `getProductsRepository` function and returning them as an array of `ProductResponseDTO`. This enhancement improves the functionality of the application by providing the ability to retrieve products based on the `companyId`.